Refactor hotel controller: Move get_hotel_controller to dependencies, update constructor to accept hotel_service
