+++ src/app/controllers/__init__.py
@@ -1,5 +1,5 @@
 """Controllers module for handling HTTP request/response logic."""
 
-from src.app.controllers.hotel_controller import HotelController, get_hotel_controller
+from src.app.controllers.hotel_controller import HotelController
 
-__all__ = ["HotelController", "get_hotel_controller"]
+__all__ = ["HotelController"]
