+++ src/app/controllers/hotel_controller.py
@@ -6,6 +6,7 @@ from pydantic import ValidationError
 
 from src.app.schemas.hotel_search import HotelSearchRequest, HotelSearchResponse
 from src.app.services.factory import ServiceFactory
+from src.app.services.hotel_service import HotelService
 from src.app.services.searchapi.client import (
     SearchAPIAuthenticationError,
     SearchAPIError,
@@ -22,9 +23,9 @@ class HotelController:
     and manages error handling with appropriate HTTP status codes.
     """
 
-    def __init__(self) -> None:
+    def __init__(self, hotel_service: HotelService | None = None) -> None:
         """Initialize hotel controller."""
-        pass
+        self.hotel_service = hotel_service or ServiceFactory.get_hotel_service()
 
     async def search_hotels(self, request: HotelSearchRequest) -> HotelSearchResponse:
         """
@@ -63,57 +64,34 @@ class HotelController:
         )
 
         try:
-            # Get hotel service instance from factory
-            hotel_service = ServiceFactory.get_hotel_service()
-
-            # Delegate to service layer for business logic
-            response = await hotel_service.search_hotels(request)
-
+            response = await self.hotel_service.search_hotels(request)
             logger.info(
                 "Hotel search completed successfully",
-                extra={
-                    "location": request.q,
-                    "results_count": len(response.properties),
-                },
+                extra={"location": request.q, "results_count": len(response.properties)},
             )
-
             return response
 
         except ValidationError as e:
-            logger.warning(
-                "Invalid request parameters",
-                extra={"location": request.q, "validation_errors": str(e)},
-            )
+            logger.warning("Invalid request parameters", extra={"location": request.q, "validation_errors": str(e)})
             raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=f"Invalid request parameters: {e}",
+                status_code=status.HTTP_400_BAD_REQUEST, detail=f"Invalid request parameters: {e}"
             ) from e
 
         except SearchAPIAuthenticationError as e:
-            logger.error(
-                "SearchAPI authentication failed",
-                extra={"location": request.q, "error": str(e)},
-            )
+            logger.error("SearchAPI authentication failed", extra={"location": request.q, "error": str(e)})
             raise HTTPException(
                 status_code=status.HTTP_401_UNAUTHORIZED,
                 detail="Authentication failed with external hotel search service",
             ) from e
 
         except SearchAPIRateLimitError as e:
-            logger.warning(
-                "SearchAPI rate limit exceeded",
-                extra={"location": request.q, "error": str(e)},
-            )
+            logger.warning("SearchAPI rate limit exceeded", extra={"location": request.q, "error": str(e)})
             raise HTTPException(
-                status_code=status.HTTP_429_TOO_MANY_REQUESTS,
-                detail="Rate limit exceeded. Please try again later.",
+                status_code=status.HTTP_429_TOO_MANY_REQUESTS, detail="Rate limit exceeded. Please try again later."
             ) from e
 
         except SearchAPITimeoutError as e:
-            logger.error(
-                "SearchAPI request timed out",
-                extra={"location": request.q, "error": str(e)},
-            )
+            logger.error("SearchAPI request timed out", extra={"location": request.q, "error": str(e)})
             raise HTTPException(
                 status_code=status.HTTP_504_GATEWAY_TIMEOUT,
                 detail="Hotel search service request timed out. Please try again.",
@@ -125,34 +103,15 @@ class HotelController:
                 extra={"location": request.q, "error": str(e), "error_type": type(e).__name__},
             )
             raise HTTPException(
-                status_code=status.HTTP_502_BAD_GATEWAY,
-                detail="External hotel search service is currently unavailable",
+                status_code=status.HTTP_502_BAD_GATEWAY, detail="External hotel search service is currently unavailable"
             ) from e
 
         except Exception as e:
             logger.error(
                 "Unexpected error during hotel search",
-                extra={
-                    "location": request.q,
-                    "error": str(e),
-                    "error_type": type(e).__name__,
-                },
+                extra={"location": request.q, "error": str(e), "error_type": type(e).__name__},
             )
             raise HTTPException(
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                 detail="An unexpected error occurred. Please try again later.",
             ) from e
-
-
-def get_hotel_controller() -> HotelController:
-    """
-    Create HotelController instance.
-
-    This function provides a simple way to get a controller instance
-    for dependency injection in FastAPI routes.
-
-    Returns:
-        HotelController: New controller instance
-
-    """
-    return HotelController()
